Migrate datatreee assertions/extensions/formatting (#8967)
* DAS-2067 - Migrate formatting.py.

* DAS-2067 - Migrate datatree/extensions.py.

* DAS-2067 - Migrate datatree/tests/test_dataset_api.py.

* DAS-2067 - Migrate datatree_render.py.

* DAS-2067 - Migrate DataTree assertions into xarray/testing/assertions.py.

* DAS-2067 - Update doc/whats-new.rst.

* DAS-2067 - Fix doctests for DataTreeRender.by_attr.

* DAS-2067 - Fix comments in doctests examples for datatree_render.

* DAS-2067 - Implement PR feedback, fix RenderDataTree.__str__.

* DAS-2067 - Add overload for xarray.testing.assert_equal and xarray.testing.assert_identical.

* DAS-2067 - Remove out-of-date comments.

* Remove test of printing datatree

---------
