+++ xarray/tests/test_backends_datatree.py
@@ -5,7 +5,7 @@ from typing import TYPE_CHECKING
 import pytest
 
 from xarray.backends.api import open_datatree
-from xarray.datatree_.datatree.testing import assert_equal
+from xarray.testing import assert_equal
 from xarray.tests import (
     requires_h5netcdf,
     requires_netCDF4,
