+++ xarray/tests/test_extensions.py
@@ -5,9 +5,14 @@ import pickle
 import pytest
 
 import xarray as xr
+
+# TODO: Remove imports in favour of xr.DataTree etc, once part of public API
+from xarray.core.datatree import DataTree
+from xarray.core.extensions import register_datatree_accessor
 from xarray.tests import assert_identical
 
 
+@register_datatree_accessor("example_accessor")
 @xr.register_dataset_accessor("example_accessor")
 @xr.register_dataarray_accessor("example_accessor")
 class ExampleAccessor:
@@ -19,6 +24,7 @@ class ExampleAccessor:
 
 class TestAccessor:
     def test_register(self) -> None:
+        @register_datatree_accessor("demo")
         @xr.register_dataset_accessor("demo")
         @xr.register_dataarray_accessor("demo")
         class DemoAccessor:
@@ -30,6 +36,9 @@ class TestAccessor:
             @property
             def foo(self):
                 return "bar"
+
+        dt: DataTree = DataTree()
+        assert dt.demo.foo == "bar"
 
         ds = xr.Dataset()
         assert ds.demo.foo == "bar"
